change integers to size_t where applicable. TODO check combine reverse edges for possible optional usage

+++ include/datastructure/RailwayNetwork.hpp
@@ -40,16 +40,16 @@ struct Edge {
    * @param breakable Boolean indicating if VSS can be placed on this edge
    * @param min_block_length Minimum block length (in m). Optional, default 1.
    */
-  int    source;
-  int    target;
+  size_t source;
+  size_t target;
   double length;
   double max_speed;
   bool   breakable;
   double min_block_length = 1;
 
   // Constructors
-  Edge(int source, int target, double length, double max_speed, bool breakable,
-       double min_block_length = 1)
+  Edge(size_t source, size_t target, double length, double max_speed,
+       bool breakable, double min_block_length = 1)
       : source(source), target(target), length(length), max_speed(max_speed),
         breakable(breakable), min_block_length(min_block_length){};
 };
@@ -60,10 +60,10 @@ class Network {
    *
    */
 private:
-  std::vector<Vertex>                  vertices;
-  std::vector<Edge>                    edges;
-  std::vector<std::vector<int>>        successors;
-  std::unordered_map<std::string, int> vertex_name_to_index;
+  std::vector<Vertex>                     vertices;
+  std::vector<Edge>                       edges;
+  std::vector<std::vector<size_t>>        successors;
+  std::unordered_map<std::string, size_t> vertex_name_to_index;
 
   void        read_graphml(const std::filesystem::path& p);
   static void get_keys(tinyxml2::XMLElement* graphml_body,
@@ -87,26 +87,26 @@ private:
   void export_successors_cpp(const std::filesystem::path& p) const;
   void write_successor_set_to_file(std::ofstream& file, int i) const;
 
-  std::pair<std::vector<int>, std::vector<int>>
-  separate_edge_at(int                        edge_index,
+  std::pair<std::vector<size_t>, std::vector<size_t>>
+  separate_edge_at(size_t                     edge_index,
                    const std::vector<double>& distances_from_source);
-  std::pair<std::vector<int>, std::vector<int>>
-  uniform_separate_edge(int edge_index);
+  std::pair<std::vector<size_t>, std::vector<size_t>>
+  uniform_separate_edge(size_t edge_index);
   // std::pair<std::vector<int>, std::vector<int>> chebychev_separate_edge(int
   // edge_index);
 
   // helper function
-  void dfs(std::vector<std::vector<int>>& ret_val,
-           std::unordered_set<int>&       vertices_to_visit,
-           const cda_rail::VertexType&    section_type) const {
+  void dfs(std::vector<std::vector<size_t>>& ret_val,
+           std::unordered_set<size_t>&       vertices_to_visit,
+           const cda_rail::VertexType&       section_type) const {
     dfs(ret_val, vertices_to_visit, section_type, {});
   };
-  void dfs(std::vector<std::vector<int>>&           ret_val,
-           std::unordered_set<int>&                 vertices_to_visit,
+  void dfs(std::vector<std::vector<size_t>>&        ret_val,
+           std::unordered_set<size_t>&              vertices_to_visit,
            const cda_rail::VertexType&              section_type,
            const std::vector<cda_rail::VertexType>& error_types) const;
-  std::vector<std::pair<int, int>>
-  sort_edge_pairs(std::vector<std::pair<int, int>>& edge_pairs) const;
+  std::vector<std::pair<size_t, size_t>>
+  sort_edge_pairs(std::vector<std::pair<size_t, size_t>>& edge_pairs) const;
 
 public:
   // Constructors
@@ -129,22 +129,23 @@ public:
   };
   [[nodiscard]] const std::vector<Edge>& get_edges() const { return edges; };
 
-  [[nodiscard]] std::vector<int>
+  [[nodiscard]] std::vector<size_t>
   get_vertices_by_type(cda_rail::VertexType type) const;
 
-  int add_vertex(const std::string& name, VertexType type);
-  int add_edge(int source, int target, double length, double max_speed,
-               bool breakable, double min_block_length = 1);
-  int add_edge(const std::string& source_name, const std::string& target_name,
-               double length, double max_speed, bool breakable,
-               double min_block_length = 1) {
+  size_t add_vertex(const std::string& name, VertexType type);
+  size_t add_edge(size_t source, size_t target, double length, double max_speed,
+                  bool breakable, double min_block_length = 1);
+  size_t add_edge(const std::string& source_name,
+                  const std::string& target_name, double length,
+                  double max_speed, bool breakable,
+                  double min_block_length = 1) {
     return add_edge(get_vertex_index(source_name),
                     get_vertex_index(target_name), length, max_speed, breakable,
                     min_block_length);
   };
-  void add_successor(int edge_in, int edge_out);
-  void add_successor(const std::pair<int, int>& edge_in,
-                     const std::pair<int, int>& edge_out) {
+  void add_successor(size_t edge_in, size_t edge_out);
+  void add_successor(const std::pair<size_t, size_t>& edge_in,
+                     const std::pair<size_t, size_t>& edge_out) {
     add_successor(get_edge_index(edge_in.first, edge_in.second),
                   get_edge_index(edge_out.first, edge_out.second));
   };
@@ -153,77 +154,79 @@ public:
     add_successor(get_edge_index(edge_in.first, edge_in.second),
                   get_edge_index(edge_out.first, edge_out.second));
   };
-  void add_successor(int edge_in, const std::pair<int, int>& edge_out) {
+  void add_successor(size_t                           edge_in,
+                     const std::pair<size_t, size_t>& edge_out) {
     add_successor(edge_in, get_edge_index(edge_out.first, edge_out.second));
   };
-  void add_successor(int                                        edge_in,
+  void add_successor(size_t                                     edge_in,
                      const std::pair<std::string, std::string>& edge_out) {
     add_successor(edge_in, get_edge_index(edge_out.first, edge_out.second));
   };
-  void add_successor(const std::pair<int, int>& edge_in, int edge_out) {
+  void add_successor(const std::pair<size_t, size_t>& edge_in,
+                     size_t                           edge_out) {
     add_successor(get_edge_index(edge_in.first, edge_in.second), edge_out);
   };
-  void add_successor(const std::pair<int, int>&                 edge_in,
+  void add_successor(const std::pair<size_t, size_t>&           edge_in,
                      const std::pair<std::string, std::string>& edge_out) {
     add_successor(get_edge_index(edge_in.first, edge_in.second),
                   get_edge_index(edge_out.first, edge_out.second));
   };
   void add_successor(const std::pair<std::string, std::string>& edge_in,
-                     int                                        edge_out) {
+                     size_t                                     edge_out) {
     add_successor(get_edge_index(edge_in.first, edge_in.second), edge_out);
   };
   void add_successor(const std::pair<std::string, std::string>& edge_in,
-                     const std::pair<int, int>&                 edge_out) {
+                     const std::pair<size_t, size_t>&           edge_out) {
     add_successor(get_edge_index(edge_in.first, edge_in.second),
                   get_edge_index(edge_out.first, edge_out.second));
   };
 
-  [[nodiscard]] const Vertex& get_vertex(int index) const;
+  [[nodiscard]] const Vertex& get_vertex(size_t index) const;
   [[nodiscard]] const Vertex& get_vertex(const std::string& name) const {
     return vertices[get_vertex_index(name)];
   };
-  [[nodiscard]] int         get_vertex_index(const std::string& name) const;
-  [[nodiscard]] const Edge& get_edge(int index) const;
-  [[nodiscard]] const Edge& get_edge(int source_id, int target_id) const;
+  [[nodiscard]] size_t      get_vertex_index(const std::string& name) const;
+  [[nodiscard]] const Edge& get_edge(size_t index) const;
+  [[nodiscard]] const Edge& get_edge(size_t source_id, size_t target_id) const;
   [[nodiscard]] const Edge& get_edge(const std::string& source_name,
                                      const std::string& target_name) const {
     return get_edge(get_vertex_index(source_name),
                     get_vertex_index(target_name));
   };
-  [[nodiscard]] int get_edge_index(int source_id, int target_id) const;
-  [[nodiscard]] int get_edge_index(const std::string& source_name,
-                                   const std::string& target_name) const {
+  [[nodiscard]] size_t get_edge_index(size_t source_id, size_t target_id) const;
+  [[nodiscard]] size_t get_edge_index(const std::string& source_name,
+                                      const std::string& target_name) const {
     return get_edge_index(get_vertex_index(source_name),
                           get_vertex_index(target_name));
   };
 
-  [[nodiscard]] bool has_vertex(int index) const {
+  [[nodiscard]] bool has_vertex(size_t index) const {
     return (index >= 0 && index < vertices.size());
   };
   [[nodiscard]] bool has_vertex(const std::string& name) const {
     return vertex_name_to_index.find(name) != vertex_name_to_index.end();
   };
-  [[nodiscard]] bool has_edge(int index) const {
+  [[nodiscard]] bool has_edge(size_t index) const {
     return (index >= 0 && index < edges.size());
   };
-  [[nodiscard]] bool has_edge(int source_id, int target_id) const;
+  [[nodiscard]] bool has_edge(size_t source_id, size_t target_id) const;
   [[nodiscard]] bool has_edge(const std::string& source_name,
                               const std::string& target_name) const;
 
-  void change_vertex_name(int index, const std::string& new_name);
+  void change_vertex_name(size_t index, const std::string& new_name);
   void change_vertex_name(const std::string& old_name,
                           const std::string& new_name) {
     change_vertex_name(get_vertex_index(old_name), new_name);
   };
-  void change_vertex_type(int index, cda_rail::VertexType new_type);
+  void change_vertex_type(size_t index, cda_rail::VertexType new_type);
   void change_vertex_type(const std::string&   name,
                           cda_rail::VertexType new_type) {
     change_vertex_type(get_vertex_index(name), new_type);
   };
 
-  void change_edge_property(int index, double value,
+  void change_edge_property(size_t index, double value,
                             const std::string& property);
-  void change_edge_property(int source_id, int target_id, double value,
+  void change_edge_property(size_t source_id, size_t target_id, double value,
                             const std::string& property) {
     change_edge_property(get_edge_index(source_id, target_id), value, property);
   };
@@ -233,8 +236,8 @@ public:
     change_edge_property(get_edge_index(source_name, target_name), value,
                          property);
   };
-  void change_edge_breakable(int index, bool value);
-  void change_edge_breakable(int source_id, int target_id, bool value) {
+  void change_edge_breakable(size_t index, bool value);
+  void change_edge_breakable(size_t source_id, size_t target_id, bool value) {
     change_edge_breakable(get_edge_index(source_id, target_id), value);
   };
   void change_edge_breakable(const std::string& source_name,
@@ -242,35 +245,35 @@ public:
     change_edge_breakable(get_edge_index(source_name, target_name), value);
   };
 
-  [[nodiscard]] std::vector<int> out_edges(int index) const;
-  [[nodiscard]] std::vector<int> out_edges(const std::string& name) const {
+  [[nodiscard]] std::vector<size_t> out_edges(size_t index) const;
+  [[nodiscard]] std::vector<size_t> out_edges(const std::string& name) const {
     return out_edges(get_vertex_index(name));
   };
-  [[nodiscard]] std::vector<int> in_edges(int index) const;
-  [[nodiscard]] std::vector<int> in_edges(const std::string& name) const {
+  [[nodiscard]] std::vector<size_t> in_edges(size_t index) const;
+  [[nodiscard]] std::vector<size_t> in_edges(const std::string& name) const {
     return in_edges(get_vertex_index(name));
   };
-  [[nodiscard]] const std::vector<int>& get_successors(int index) const;
-  [[nodiscard]] const std::vector<int>& get_successors(int source_id,
-                                                       int target_id) const {
+  [[nodiscard]] const std::vector<size_t>& get_successors(size_t index) const;
+  [[nodiscard]] const std::vector<size_t>&
+  get_successors(size_t source_id, size_t target_id) const {
     return get_successors(get_edge_index(source_id, target_id));
   };
-  [[nodiscard]] const std::vector<int>&
+  [[nodiscard]] const std::vector<size_t>&
   get_successors(const std::string& source_name,
                  const std::string& target_name) const {
     return get_successors(get_edge_index(source_name, target_name));
   };
 
-  [[nodiscard]] std::vector<int> neighbors(int index) const;
-  [[nodiscard]] std::vector<int> neighbors(const std::string& name) const {
+  [[nodiscard]] std::vector<size_t> neighbors(size_t index) const;
+  [[nodiscard]] std::vector<size_t> neighbors(const std::string& name) const {
     return neighbors(get_vertex_index(name));
   };
 
   [[nodiscard]] size_t number_of_vertices() const { return vertices.size(); };
   [[nodiscard]] size_t number_of_edges() const { return edges.size(); };
 
-  [[nodiscard]] int max_vss_on_edge(int index) const;
-  [[nodiscard]] int max_vss_on_edge(int source, int target) const {
+  [[nodiscard]] int max_vss_on_edge(size_t index) const;
+  [[nodiscard]] int max_vss_on_edge(size_t source, size_t target) const {
     return max_vss_on_edge(get_edge_index(source, target));
   };
   [[nodiscard]] int max_vss_on_edge(const std::string& source,
@@ -295,9 +298,9 @@ public:
   };
   void export_network(const std::filesystem::path& p) const;
 
-  [[nodiscard]] bool is_valid_successor(int e0, int e1) const;
+  [[nodiscard]] bool is_valid_successor(size_t e0, size_t e1) const;
 
-  [[nodiscard]] bool is_adjustable(int vertex_id) const;
+  [[nodiscard]] bool is_adjustable(size_t vertex_id) const;
   [[nodiscard]] bool is_adjustable(const std::string& vertex_name) const {
     return is_adjustable(get_vertex_index(vertex_name));
   };
@@ -305,40 +308,40 @@ public:
   [[nodiscard]] bool is_consistent_for_transformation() const;
 
   // Get special edges
-  [[nodiscard]] std::vector<int>              breakable_edges() const;
-  [[nodiscard]] std::vector<int>              relevant_breakable_edges() const;
-  [[nodiscard]] std::vector<std::vector<int>> unbreakable_sections() const;
-  [[nodiscard]] std::vector<std::vector<int>> no_border_vss_sections() const;
-  [[nodiscard]] std::vector<std::pair<int, int>>
-  combine_reverse_edges(const std::vector<int>& edges_to_consider,
-                        bool                    sort = false) const;
-  [[nodiscard]] int get_reverse_edge_index(int edge_index) const;
-  [[nodiscard]] std::optional<int>
-  common_vertex(const std::pair<int, int>& pair1,
-                const std::pair<int, int>& pair2) const;
+  [[nodiscard]] std::vector<size_t> breakable_edges() const;
+  [[nodiscard]] std::vector<size_t> relevant_breakable_edges() const;
+  [[nodiscard]] std::vector<std::vector<size_t>> unbreakable_sections() const;
+  [[nodiscard]] std::vector<std::vector<size_t>> no_border_vss_sections() const;
+  [[nodiscard]] std::vector<std::pair<size_t, size_t>>
+  combine_reverse_edges(const std::vector<size_t>& edges_to_consider,
+                        bool                       sort = false) const;
+  [[nodiscard]] size_t get_reverse_edge_index(size_t edge_index) const;
+  [[nodiscard]] std::optional<size_t>
+  common_vertex(const std::pair<size_t, size_t>& pair1,
+                const std::pair<size_t, size_t>& pair2) const;
 
-  [[nodiscard]] std::vector<int>
-  inverse_edges(const std::vector<int>& edge_indices) const {
-    const auto&      edge_number = number_of_edges();
-    std::vector<int> edges_to_consider(edge_number);
+  [[nodiscard]] std::vector<size_t>
+  inverse_edges(const std::vector<size_t>& edge_indices) const {
+    const auto&         edge_number = number_of_edges();
+    std::vector<size_t> edges_to_consider(edge_number);
     std::iota(edges_to_consider.begin(), edges_to_consider.end(), 0);
     return inverse_edges(edge_indices, edges_to_consider);
   };
-  [[nodiscard]] std::vector<int>
-  inverse_edges(const std::vector<int>& edge_indices,
-                const std::vector<int>& edges_to_consider) const;
+  [[nodiscard]] std::vector<size_t>
+  inverse_edges(const std::vector<size_t>& edge_indices,
+                const std::vector<size_t>& edges_to_consider) const;
 
   // Transformation functions
-  std::pair<std::vector<int>, std::vector<int>>
-  separate_edge(int edge_index, cda_rail::SeparationType separation_type =
-                                    cda_rail::SeparationType::UNIFORM);
-  std::pair<std::vector<int>, std::vector<int>>
-  separate_edge(int source_id, int target_id,
+  std::pair<std::vector<size_t>, std::vector<size_t>>
+  separate_edge(size_t edge_index, cda_rail::SeparationType separation_type =
+                                       cda_rail::SeparationType::UNIFORM);
+  std::pair<std::vector<size_t>, std::vector<size_t>>
+  separate_edge(size_t source_id, size_t target_id,
                 cda_rail::SeparationType separation_type =
                     cda_rail::SeparationType::UNIFORM) {
     return separate_edge(get_edge_index(source_id, target_id), separation_type);
   };
-  std::pair<std::vector<int>, std::vector<int>>
+  std::pair<std::vector<size_t>, std::vector<size_t>>
   separate_edge(const std::string& source_name, const std::string& target_name,
                 cda_rail::SeparationType separation_type =
                     cda_rail::SeparationType::UNIFORM) {
@@ -346,7 +349,7 @@ public:
                          separation_type);
   };
 
-  std::vector<std::pair<int, std::vector<int>>>
+  std::vector<std::pair<size_t, std::vector<size_t>>>
   discretize(cda_rail::SeparationType separation_type =
                  cda_rail::SeparationType::UNIFORM);
